requirements.txt
# Agno framework (main dependency)
agno>=1.7.6

# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
openai>=1.12.0
psycopg2-binary>=2.9.5
supabase>=2.3.0
pgvector>=0.2.4
tiktoken>=0.5.2
markdown-it-py>=3.0.0

# API and web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# Data processing
pandas>=2.1.4
numpy>=1.24.3
scipy>=1.11.4
scikit-learn>=1.3.2

# Utilities
pydantic>=2.5.0
loguru>=0.7.2
rich>=13.7.0
typer>=0.9.0
python-dotenv>=1.0.0
watchdog>=3.0.0
psutil>=5.9.0
